+++ src/omnisharp/engines/IEngine.ts
@@ -1,3 +1,8 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
 import { CancellationToken } from '../../vscodeAdapter';
 import { LaunchTarget } from '../launcher';
 import { LaunchInfo } from '../OmnisharpManager';
