+++ src/omnisharp/engines/StdioEngine.ts
@@ -1,3 +1,8 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
 import * as protocol from '../protocol';
 import * as utils from '../../common';
 import { CancellationToken } from '../../vscodeAdapter';
